Blacklist services that will not update without changes
Since there hasn't been a complete update of the AWS models in
quite a while, there's a backlog of services to support.

Rather than add support for them all in one commit, which may
not do due justice to each one, this allows the models to be
updated one by one with proper oversight.

At the very end of the list are a few services that currently do
not compile after update (alexaforbusiness onwards), and need
to be fixed.